자료형 변경

+++ src/main/java/com/eeerrorcode/guestbook/domain/dto/BoardDto.java
@@ -19,5 +19,5 @@ public class BoardDto {
   private String memberName;
   private LocalDateTime regDate;
   private LocalDateTime modDate;
-  private int replyCnt;
+  private Long replyCnt;
 }
